samples/HttpServer_FirmwareUpload/requirements.txt
# Requirements file for pip
# list of Python packages used in documentation build
PyNaCl
